Add files via upload
An unfinished version of the project. It has most of the meat of my website. What's left is a few media queries that will make it functional on larger tablets and laptops/desktops. I just recently changed all the links to relative paths so it can work outside of my computer.